+++ dsh_sdk/src/utils/mod.rs
@@ -12,6 +12,8 @@ use crate::error::DshError;
 pub mod dlq;
 #[cfg(feature = "graceful-shutdown")]
 pub mod graceful_shutdown;
+#[cfg(feature = "hyper-client")]
+pub(crate) mod http_client;
 #[cfg(feature = "metrics")]
 pub mod metrics;